Refactor trace writer options module
